Update README.md
This is a repository where I upload the task 1 of The Spark Foundation Internship. I have completed this task using supervised machine learning. The task is about the Predict the percentage of an student based on the no. of study hours such as What will be predicted score if a student studies for 9.25 hrs/ day?.